# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Salon Billing Software API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponse"
    post:
      operationId: createService
      tags: [services]
      summary: Create a new service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    put:
      operationId: updateService
      tags: [services]
      summary: Update service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Service deleted
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff members
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffListResponse"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a new staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
      responses:
        "201":
          description: Staff created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
  /staff/{id}:
    get:
      operationId: getStaff
      tags: [staff]
      summary: Get staff details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffDetail"
    put:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
      responses:
        "200":
          description: Staff updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
...
components:
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        price:
          type: number
        duration:
          type: integer
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
      required: [id, name, category, price, duration, isActive]
    ServiceListResponse:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        categories:
          type: array
          items:
            type: string
      required: [services, categories]
    CreateServiceRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        duration:
          type: integer
        description:
          type: string
          nullable: true
      required: [name, category, price, duration]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        category:
          type: string
        stockQuantity:
          type: integer
        reorderLevel:
          type: integer
        costPrice:
          type: number
        sellingPrice:
          type: number
        supplier:
          type: string
          nullable: true
        isLowStock:
          type: boolean
      required: [id, name, brand, category, stockQuantity, reorderLevel, costPrice, sellingPrice, isLowStock]
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
      required: [products, total]
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        category:
          type: string
        stockQuantity:
          type: integer
        reorderLevel:
          type: integer
        costPrice:
          type: number
        sellingPrice:
          type: number
        supplier:
          type: string
          nullable: true
      required: [name, brand, category, stockQuantity, reorderLevel, costPrice, sellingPrice]
    Staff:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialization:
          type: string
        commissionPercent:
          type: number
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        isActive:
          type: boolean
        workingHours:
          type: string
          nullable: true
        avatarInitials:
          type: string
      required: [id, name, specialization, commissionPercent, isActive, avatarInitials]
    StaffDetail:
      allOf:
        - $ref: "#/components/schemas/Staff"
        - type: object
          properties:
            servicesPerformed:
              type: integer
            revenueGenerated:
              type: number
            tips:
              type: number
          required: [servicesPerformed, revenueGenerated, tips]
    StaffListResponse:
      type: object
      properties:
        staff:
          type: array
          items:
            $ref: "#/components/schemas/Staff"
      required: [staff]
    CreateStaffRequest:
      type: object
      properties:
        name:
          type: string
        specialization:
          type: string
        commissionPercent:
          type: number
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        workingHours:
          type: string
          nullable: true
      required: [name, specialization, commissionPercent]
...