# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salon Billing Software API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  brand: string;
  category: string;
  stockQuantity: number;
  reorderLevel: number;
  costPrice: number;
  sellingPrice: number;
  supplier?: string | null;
  isLowStock: boolean;
}