# lib/api-zod/src/generated/types/serviceListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Salon Billing Software API
 * OpenAPI spec version: 0.1.0
 */
import type { Service } from "./service";

export interface ServiceListResponse {
  services: Service[];
  categories: string[];
}